backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
jiwer==3.0.4
openai==1.50.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
deepgram-sdk==3.7.6
assemblyai==0.32.0
